pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fyta-mcp-server"
version = "1.0.0"
description = "MCP server for FYTA plant sensor data"
authors = [
    {name = "Juergen Schilling", email = "juergen_schilling@web.de"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
fyta-mcp-server = "fyta_mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fyta_mcp_server"]
